refactor email Sender
